Crate devicemapper

Low-level devicemapper configuration of the running kernel.

Overview

Linux's devicemapper allows the creation of block devices whose storage is mapped to other block devices in useful ways, either by changing the location of its data blocks, or performing some operation on the data itself. This is a low-level facility that is used by higher-level volume managers such as LVM2. Uses may include:

Dividing a large block device into smaller logical volumes (dm-linear)

Combining several separate block devices into a single block device with better performance and/or redundancy (dm-raid)

Encrypting a block device (dm-crypt)

Performing Copy-on-Write (COW) allocation of a volume's blocks enabling fast volume cloning and snapshots (dm-thin)

Configuring a smaller, faster block device to act as a cache for a larger, slower one (dm-cache)

Verifying the contents of a read-only volume (dm-verity)

Usage

Before they can be used, DM devices must be created using DM::device_create(), have a mapping table loaded using DM::table_load(), and then activated with DM::device_resume(). Once activated, they can be used as a regular block device, including having other DM devices map to them.

Devices have "active" and "inactive" mapping tables. See function descriptions for which table they affect.

Structs

DM
Context needed for communicating with devicemapper.

Device
A struct containing the device's major and minor numbers

Enums

StatusType
Used with DM::table_status() to choose either return of info or tables for a target. The contents of each of these strings is target-specific.

Functions

dev_majors
Major numbers used by DM.